+++ sdk/communication/azure-communication-callautomation/src/main/java/com/azure/communication/callautomation/CallAutomationClientBuilder.java
@@ -377,8 +377,8 @@ public final class CallAutomationClientBuilder implements
 
         AzureCommunicationCallAutomationServiceImplBuilder clientBuilder = new AzureCommunicationCallAutomationServiceImplBuilder();
         try {
-            clientBuilder.endpoint(new URL(endpoint)).pipeline(builderPipeline);
-        } catch (MalformedURLException e) {
+            clientBuilder.endpoint(endpoint).pipeline(builderPipeline);
+        } catch (RuntimeException e) {
             throw logger.logExceptionAsError(new RuntimeException(e.getMessage()));
         }