+++ sdk/maps/azure-maps-search/src/main/java/com/azure/maps/search/implementation/models/GeoJsonPoint.java
@@ -19,13 +19,15 @@ import java.util.List;
 @Fluent
 public final class GeoJsonPoint extends GeoJsonGeometry {
     /*
-     * A `Position` is an array of numbers with two or more elements. The first
-     * two elements are _longitude_ and _latitude_, precisely in that order.
-     * _Altitude/Elevation_ is an optional third element. Please refer to [RFC
+     * A `Position` is an array of numbers with two or more elements. The first two elements are _longitude_ and
+     * _latitude_, precisely in that order. _Altitude/Elevation_ is an optional third element. Please refer to [RFC
      * 7946](https://tools.ietf.org/html/rfc7946#section-3.1.1) for details.
      */
     @JsonProperty(value = "coordinates", required = true)
     private List<Double> coordinates;
+
+    /** Creates an instance of GeoJsonPoint class. */
+    public GeoJsonPoint() {}
 
     /**
      * Get the coordinates property: A `Position` is an array of numbers with two or more elements. The first two